reset workflow with no decision task complete (#3687)

+++ service/history/execution/context.go
@@ -950,7 +950,7 @@ func (c *contextImpl) PersistNonFirstWorkflowEvents(
 			// TransactionID is set by shard context
 		},
 	)
-	return int64(size), err
+	return size, err
 }
 
 func (c *contextImpl) appendHistoryV2EventsWithRetry(
